requirements.txt
crewai
openai
langchain-openai
streamlit
pandas
duckdb
python-dotenv
tabulate
pytest
